+++ package.json
@@ -12,6 +12,7 @@
     "@caloriosa/rest-dto": "file:modules/rest-dto",
     "body-parser": "~1.18.2",
     "bootstrap": "^4.0.0-beta.2",
+    "config": "^1.28.1",
     "connect-flash": "^0.1.1",
     "connect-memcached": "^0.2.0",
     "cookie-parser": "~1.4.3",
@@ -26,6 +27,7 @@
     "log4js": "^2.3.12",
     "morgan": "~1.9.0",
     "node-sass-middleware": "^0.11.0",
+    "node-yaml-config": "0.0.4",
     "nodemon": "^1.12.5",
     "popper.js": "^1.12.9",
     "serve-favicon": "~2.4.5",
